Code_Refacto

+++ BookLibrary/Models/Book.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 
 namespace BookLibrary.Models {
-
-
     public class Book {
-        public int ISBN { get; set; }
+        public decimal ISBN { get; set; }          // numeric(13,0)
         public string Title { get; set; }
         public string Author { get; set; }
         public DateTime PublishDate { get; set; }
